Changed int to size_t in countstr()
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1541 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/CSeries/csstrings.cpp
@@ -46,7 +46,7 @@ Jan 25, 2002 (Br'fin (Jeremy Parsons)):
 char temporary[256];
 
 
-short countstr(
+size_t countstr(
 	short resid)
 {
 	return TS_CountStrings(resid);
